+++ packages/authhero/src/authentication-flows/common.ts
@@ -92,6 +92,7 @@ export async function createAuthTokens(
   if (ctx.env.hooks?.onExecuteCredentialsExchange) {
     await ctx.env.hooks.onExecuteCredentialsExchange(
       {
+        ctx,
         client,
         user,
         request: {
@@ -257,6 +258,7 @@ export async function createSession(
   const session = await ctx.env.data.sessions.create(client.tenant.id, {
     id: nanoid(),
     user_id: user.user_id,
+    login_session_id: loginSession.id,
     idle_expires_at: new Date(
       Date.now() + SILENT_AUTH_MAX_AGE_IN_SECONDS * 1000,
     ).toISOString(),