+++ packages/authhero/test/authentication-flows/silent.spec.ts
@@ -8,10 +8,21 @@ describe("silent", () => {
     const { oauthApp, env } = await getTestServer();
     const oauthClient = testClient(oauthApp, env);
 
+    const loginSession = await env.data.loginSessions.create("tenantId", {
+      expires_at: new Date(Date.now() + 3600 * 1000).toISOString(),
+      csrf_token: "csrfToken",
+      authParams: {
+        client_id: "clientId",
+        redirect_uri: "https://example.com/callback",
+        response_type: AuthorizationResponseType.TOKEN_ID_TOKEN,
+      },
+    });
+
     const session = await env.data.sessions.create("tenantId", {
       id: "sessionId",
       user_id: "email|userId",
       used_at: new Date().toISOString(),
+      login_session_id: loginSession.id, // Link session to login session
       device: {
         last_ip: "",
         initial_ip: "",
@@ -24,7 +35,10 @@ describe("silent", () => {
       clients: ["clientId"],
     });
 
-    const response = await oauthClient.authorize.$get(
+    // ----------------------------------------
+    // Get a token and id-token for the session
+    // ----------------------------------------
+    const accessTokenResponse = await oauthClient.authorize.$get(
       {
         query: {
           client_id: "clientId",
@@ -42,9 +56,44 @@ describe("silent", () => {
       },
     );
 
-    expect(response.status).toEqual(200);
-    const htmlBody = await response.text();
-    expect(htmlBody).toContain("access_token");
+    expect(accessTokenResponse.status).toEqual(200);
+    const accessTokenhtmlBody = await accessTokenResponse.text();
+    expect(accessTokenhtmlBody).toContain("access_token");
+
+    // ----------------------------------------
+    // Get a code for the session
+    // ----------------------------------------
+    const codeResponse = await oauthClient.authorize.$get(
+      {
+        query: {
+          client_id: "clientId",
+          redirect_uri: "https://example.com/callback",
+          state: "state",
+          prompt: "none",
+          response_type: AuthorizationResponseType.CODE,
+        },
+      },
+      {
+        headers: {
+          origin: "https://example.com",
+          cookie: `tenantId-auth-token=${session.id}`,
+        },
+      },
+    );
+
+    expect(codeResponse.status).toEqual(200);
+    const codeResponseHtmlBody = await codeResponse.text();
+    expect(codeResponseHtmlBody).toContain("code");
+
+    const codeMatch = codeResponseHtmlBody.match(/"code":"([^"]+)"/);
+
+    const code = await env.data.codes.get(
+      "tenantId",
+      codeMatch?.[1] || "",
+      "authorization_code",
+    );
+    expect(code).toBeDefined();
+    expect(code?.login_id).toEqual(loginSession.id);
 
     // Check that the session was updated
     const updatedSession = await env.data.sessions.get("tenantId", "sessionId");