+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @authhero/demo
 
+## 0.13.2
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.149.0
+  - @authhero/kysely-adapter@10.17.0
+
+## 0.13.1
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.148.0
+
 ## 0.13.0
 
 ### Minor Changes