+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,22 @@
 # authhero
 
+## 0.149.0
+
+### Minor Changes
+
+- pnpm changeset version
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.67.0
+
+## 0.148.0
+
+### Minor Changes
+
+- Pass the context to the events
+
 ## 0.147.0
 
 ### Minor Changes